Periodic merge from 1.5 to trunk (4834-4837)
git-svn-id: https://svn.kuali.org/repos/kpme/trunk@4843 6c5f4195-72ea-6250-e7e4-e498f552a5df

+++ tk-lm/impl/src/main/java/org/kuali/kpme/tklm/leave/accrual/service/AccrualServiceImpl.java
@@ -365,7 +365,9 @@ public class AccrualServiceImpl implements AccrualService {
 					sstoUsageList.add(lb);
 				}
 			} else {
-				LmServiceLocator.getLeaveBlockService().deleteLeaveBlock(lb.getLmLeaveBlockId(), runAsPrincipalId);
+				if(!(StringUtils.equals(lb.getLeaveBlockType(),LMConstants.LEAVE_BLOCK_TYPE.BALANCE_TRANSFER) ||
+								StringUtils.equals(lb.getLeaveBlockType(),LMConstants.LEAVE_BLOCK_TYPE.LEAVE_PAYOUT)))
+					LmServiceLocator.getLeaveBlockService().deleteLeaveBlock(lb.getLmLeaveBlockId(), runAsPrincipalId);
 			}
 		}
